requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.8.3
gunicorn==21.2.0
timezonefinder==6.2.0
pytz==2023.3